+++ Libraries/graphics/rect/rect.hpp
@@ -42,7 +42,7 @@ public:
     [[nodiscard]] inline unsigned short getHeight() const override { return h; };
     unsigned short w, h;
     Color c;
-    unsigned short blur_intensity = 0;
+    float blur_intensity = 0;
     void render(bool fill=true) const;
 };
